Tweaked colour parser to only accept unsigned ints in range 0 to 255.  This seems to work, as the parser barfs if given a number outside that range... however the parsed uints still aren't being passed into the GG::Clr object properly, as these end up with seemingly random (but consistent between Clr's) values in them when constructed by the parser, as seen in console debug output.  These Clr still aren't used for anything, though, and presumably won't until this is fixed.
svn path=/trunk/FreeOrion/; revision=2333

+++ universe/Parser.h
@@ -63,7 +63,7 @@ struct NameClosure : boost::spirit::closure<NameClosure, std::string>
     member1 this_;
 };
 
-struct ColourClosure : boost::spirit::closure<ColourClosure, GG::Clr, GLubyte, GLubyte, GLubyte, GLubyte>
+struct ColourClosure : boost::spirit::closure<ColourClosure, GG::Clr, unsigned int, unsigned int, unsigned int, unsigned int>
 {
     member1 this_;
     member2 r;
